backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
pygraphviz==2.0b0.dev0
PyJWT==2.9.0
razorpay==1.4.2
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.6
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
